Renames Php Tips to PhpStorm tips
